+++ packages/sdk/src/server/index.ts
@@ -17,6 +17,11 @@ export type OAuthCredentials = {
 };
 
 /**
+ * The environment in which the server client is running.
+ */
+export type ProjectEnvironment = "development" | "production";
+
+/**
  * Options for creating a server-side client.
  * This is used to configure the BackendClient instance.
  */
@@ -25,7 +30,7 @@ export type BackendClientOpts = {
    * The environment in which the server client is running (e.g., "production",
    * "development").
    */
-  environment?: string;
+  environment?: ProjectEnvironment;
 
   /**
    * The credentials to use for authentication against the Pipedream API.
@@ -353,7 +358,8 @@ export class BackendClient {
    * @param opts - The options for configuring the server client.
    */
   constructor(opts: BackendClientOpts) {
-    this.environment = opts.environment ?? "production";
+    this.ensureValidEnvironment(opts.environment);
+    this.environment = opts.environment!;
 
     this.projectId = opts.projectId;
     if (!this.projectId) {
@@ -368,6 +374,15 @@ export class BackendClient {
     this.workflowDomain = workflowDomain;
 
     this.oauthClient = this.newOauthClient(opts.credentials, this.baseApiUrl);
+  }
+
+  private ensureValidEnvironment(environment?: string) {
+    if (!environment || ![
+      "development",
+      "production",
+    ].includes(environment)) {
+      throw new Error("Project environment is required. Supported environments are development and production.");
+    }
   }
 
   private newOauthClient(