+++ packages/sdk/CHANGELOG.md
@@ -1,12 +1,21 @@
 <!-- markdownlint-disable MD024 -->
 # Changelog
 
+## [1.0.5] - 2024-11-18
+
+### Changed
+
+- The backend client used to default to `production` if the environment was not
+specified. Now `environment` is a required argument for `createBackendClient`
+and must be one of `production` or `development`.  
+
 ## [1.0.4] - 2024-11-15
 
 ### Changed
 
 - Improved the docs of the `getAccountById` method in the backend client to
   clarify the behavior of the new argument.
+
 - Fixed the exported `HTTPAuthType` enum so that it can be used by the consumers
   of the SDK.